Created and Implemented the Whiteboard, Marker and eraser
Ive added a compressed unitypackage of a whiteboard, Marker, Eraser, scene and all models needed to add the whiteboard and current features into the main project. 

The whiteboard Currently sets its own texture and changes it colour.

The marker finds its location on the Whiteboard texture, calculates how many pixels need to change based on a editor set value and then changes the pixels. It also interpolates the pixels between two drawn points if its been moved too fast.

The Eraser works very similarly except its casts two points, one in the top left and the other in the bottom right, calculates the area needing to be erased, changes the pixels underneath the eraser to the original colour of the whiteboard. 

The scene is just a floating whiteboard with an eraser and marker to test. 

All models and materials are in their own folder